requirements.txt
Django==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.5.0
drf-nested-routers==0.93.5
drf-yasg==1.21.7
PyJWT==2.9.0 